Add generate-authors.sh script
This script automates the generation of an AUTHORS file for the repository. It retrieves the commit history using git log, formats each commit with the author's name and email, sorts the list of authors, and removes duplicates. The resulting list is then written to the AUTHORS file.

The script is designed to be run from the root directory of the repository and outputs the AUTHORS file in the same directory.

Usage:
./generate-authors.sh

This script simplifies the process of crediting all individuals who have contributed content to the repository.